+++ R/define_init_values.R
@@ -18,18 +18,18 @@
 #'   continuous? If \code{TRUE}, an outcome precision parameter is initialized;
 #'   if \code{FALSE}, no outcome precision is included.
 #'
-#' @param prec.m Numeric or \code{NULL}. Initial precision (inverse variance)
+#' @param m.prec.init Numeric or \code{NULL}. Initial precision (inverse variance)
 #'   of the mediator \code{M} when \code{M_cont = TRUE}. If \code{NULL},
 #'   a default value of  is used internally.
-#' @param prec.y Numeric or \code{NULL}. Initial precision (inverse variance)
+#' @param y.prec.init Numeric or \code{NULL}. Initial precision (inverse variance)
 #'   of the outcome variable \code{Y} when \code{Y_cont = TRUE}. If \code{NULL},
 #'   a default value of 1 is used internally.
-#' @param c.prime Numeric or \code{NULL}. Initial value for the direct effect
+#' @param c.prime.init Numeric or \code{NULL}. Initial value for the direct effect
 #'   (\code{X -> Y}). If \code{NULL}, a default value is used internally.
-#' @param taua Numeric or \code{NULL}. Initial prior precision for the \code{a}
+#' @param taua.init Numeric or \code{NULL}. Initial prior precision for the \code{a}
 #'   path coefficients (\code{X -> M}). If \code{NULL}, a default value is used
 #'   internally.
-#' @param taub Numeric or \code{NULL}. Initial prior precision for the \code{b}
+#' @param taub.init Numeric or \code{NULL}. Initial prior precision for the \code{b}
 #'   path coefficients (\code{M -> Y}). If \code{NULL}, a default value is used
 #'   internally.
 #' @param ind.p Numeric in (0, 1) or \code{NULL}. Initial prior inclusion
@@ -40,7 +40,7 @@
 #' Elements included depend on the distributional assumptions for the mediator
 #' and outcome:
 #' \describe{
-#'   \item{prec.m}{Numeric vector of length \code{K}. Initial precision of each
+#'   \item{m.prec.init}{Numeric vector of length \code{K}. Initial precision of each
 #'     mediator (included only when \code{M_cont = TRUE}).}
 #'   \item{aI}{Numeric matrix of dimension \code{K × P}. Initial values for
 #'     \code{a} path coefficients.}
@@ -50,12 +50,13 @@
 #'     path coefficients.}
 #'   \item{ind.b}{Binary vector of length \code{K}. Inclusion indicators for
 #'     \code{b} paths.}
-#'   \item{prec.y}{Numeric. Outcome precision (included only when
+#'   \item{y.prec.init}{Numeric. Outcome precision (included only when
 #'     \code{Y_cont = TRUE}).}
-#'   \item{c.prime}{Numeric vector of length \code{P}. Direct effect coefficients.}
-#'   \item{taua}{Numeric. Prior precision for \code{a} paths.}
-#'   \item{taub}{Numeric. Prior precision for \code{b} paths.}
-#'   \item{ind.p}{Numeric. Prior inclusion probability.}
+#'   \item{c.prime.init}{Numeric vector of length \code{P}. Direct effect coefficients.}
+#'   \item{taua.init}{Numeric. Prior precision for \code{a} paths.}
+#'   \item{taub.init}{Numeric. Prior precision for \code{b} paths.}
+#'   \item{a.pip.init}{Numeric. Prior inclusion probability for A effect.}
+#'   \item{b.pip.init}{Numeric. Prior inclusion probability for B effect.}
 #' }
 #'
 #' @seealso \code{\link{fit_ebmed}}
@@ -68,40 +69,42 @@ define_init_values <- function(
     K,
     M_cont,
     Y_cont,
-    prec.m,
-    prec.y,
-    c.prime,
-    taua,
-    taub,
-    ind.p
+    m.prec.init,
+    y.prec.init,
+    c.prime.init,
+    taua.init,
+    taub.init,
+    a.pip.init,
+    b.pip.init
 ) {
   # Handling null cases explicitly and set values as default if NULL
-  if (is.null(prec.m)) prec.m <- 1
-  if (is.null(prec.y)) prec.y <- 1
-  if (is.null(c.prime)) c.prime <- 0
-  if (is.null(taua)) taua <- 1
-  if (is.null(taub)) taub <- 1
-  if (is.null(ind.p)) ind.p <- 0.5
+  if (is.null(m.prec.init)) m.prec.init <- 1
+  if (is.null(y.prec.init)) y.prec.init <- 1
+  if (is.null(c.prime.init)) c.prime.init <- 0
+  if (is.null(taua.init)) taua.init <- 1
+  if (is.null(taub.init)) taub.init <- 1
+  if (is.null(a.pip.init)) a.pip.init <- 0.5
+  if (is.null(b.pip.init)) b.pip.init <- 0.5
 
   #Actual Code
 
   c(
-    if (M_cont) list(prec.m = rep(prec.m, K)) else list(),
+    if (M_cont) list(m.prec = rep(m.prec.init, K)) else list(),
 
     list(
-      aI = matrix(0, nrow = K, ncol = P),
-      ind.a = rep(0, K),
-      bI = rep(0, K),
-      ind.b = rep(0, K)
+      a.coef = matrix(0, nrow = K, ncol = P),
+      a.pip = rep(0, K),
+      b.coef = rep(0, K),
+      b.pip = rep(0, K)
     ),
 
-    if (Y_cont) list(prec.y = prec.y) else list(),
+    if (Y_cont) list(y.prec = y.prec.init) else list(),
 
     list(
-      c.prime = rep(c.prime, P),
-      taua = taua,
-      taub = taub,
-      ind.p = ind.p
+      direct.coef = rep(c.prime.init, P),
+      a.coef.hyperprec = taua.init,
+      b.coef.hyperprec = taub.init,
+      ind.p = a.pip.init
     )
   )
 }