+++ tests/testthat/test-buzzMcontYcat.R
@@ -1,14 +1,14 @@
-test_that("buzzMYcat returns an mcmc.list object", {
+test_that("buzzMcontYcat returns an mcmc.list object", {
   # 1. Create minimal test data
   set.seed(2026)
   test_dataset <- data.frame(
-    outcome = rnorm(20),
-    predictor = rnorm(20),
+    outcome = rbinom(20, 1, 0.5),
+    predictor = rbinom(20, 1, 0.5),
     mediator1 = rnorm(20)
   )
 
   # 2. Run the function with minimal iterations
-  res <- buzzMYcont(
+  res <- buzzMcontYcat(
     dataset = test_dataset,
     X = "predictor",
     Y = "outcome",