+++ NEWS.md
@@ -1,3 +1,10 @@
+# buzzMed 0.1.2
+* Renamed the following functions for clarification:
+  * buzzMcat -> buzzMcatYcont
+  * buzzMYcat -> buzzMcatYcat
+  * buzzMYcont -> buzzMcontYcont
+  * buzzYcat -> buzzMcontYcat
+
 # buzzMed 0.1.1
 
 * Added unit tests using `testthat` to ensure model stability.